// frontend/src/app/models/pronouns.ts
export class Pronouns {
  infinitiv?: string;

  preschentsing1?: string;
  preschentsing2?: string;
  preschentsing3?: string;
  preschentplural1?: string;
  preschentplural2?: string;
  preschentplural3?: string;

  imperfectsing1?: string;
  imperfectsing2?: string;
  imperfectsing3?: string;
  imperfectplural1?: string;
  imperfectplural2?: string;
  imperfectplural3?: string;

  participperfectms?: string;
  participperfectfs?: string;
  participperfectmp?: string;
  participperfectfp?: string;

  conjunctivsing1?: string;
  conjunctivsing2?: string;
  conjunctivsing3?: string;
  conjunctivplural1?: string;
  conjunctivplural2?: string;
  conjunctivplural3?: string;

  conjunctiv2sing1?: string;
  conjunctiv2sing2?: string;
  conjunctiv2sing3?: string;
  conjunctiv2plural1?: string;
  conjunctiv2plural2?: string;
  conjunctiv2plural3?: string;

  cundizionalsing1?: string;
  cundizionalsing2?: string;
  cundizionalsing3?: string;
  cundizionalplural1?: string;
  cundizionalplural2?: string;
  cundizionalplural3?: string;

  cundizionalindirectsing1?: string;
  cundizionalindirectsing2?: string;
  cundizionalindirectsing3?: string;
  cundizionalindirectplural1?: string;
  cundizionalindirectplural2?: string;
  cundizionalindirectplural3?: string;

  futursing1?: string;
  futursing2?: string;
  futursing3?: string;
  futurplural1?: string;
  futurplural2?: string;
  futurplural3?: string;

  futurdubitativsing1?: string;
  futurdubitativsing2?: string;
  futurdubitativsing3?: string;
  futurdubitativplural1?: string;
  futurdubitativplural2?: string;
  futurdubitativplural3?: string;

  imperativ1?: string;
  imperativ2?: string;

  gerundium?: string;
}